warehouse_tests/requirements.txt
requests
pytest
pytest-html
pyyaml
pyodbc